requirements.txt
# Core Web Framework and UI
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
aiohttp>=3.8.0

# Database and Storage
neo4j>=5.0.0
pinecone>=2.0.0
chromadb>=0.4.0

# AI/ML Core
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
timm>=0.9.0
sentence-transformers>=2.2.0

# Computer Vision and OCR
opencv-python>=4.8.0
pillow>=10.0.0
easyocr>=1.7.0
paddleocr>=2.7.0
ultralytics>=8.0.0
supervision>=0.16.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0

# Python Core Extensions
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=4.9.0
typing-extensions>=4.7.0

# Android/Mobile Automation
pure-python-adb>=0.3.0.dev0

# Utilities (no installation needed - built-in)
# asyncio, base64, dataclasses, datetime, enum, hashlib, io, json, os, pathlib, 
# queue, re, subprocess, sys, tempfile, threading, time, uuid, ast, typing

# Optional/Development
jax>=0.4.0  # Optional for some ML models